Fixing Issue #409: add dynamic sizing to facets To resolve the issue of long collection names wrapping, then the facet will utilize a different style class so the text is smaller and will appear in a single line.  The facet value length will be compared to determine if the custom class will be utilized or not.
Added similar logic to the chiclet since the previous chiclets
were not being displayed properly, all the UI would show was
"Collection:" and you would not be able to see any text.  Added
hover over tooltip for the chiclet values so that the user can
see the full facet name.