Wrapped display of route information when clicking on a route in header
I won't be able to make improvements over this code, as I am out of time.  It is functional, but not elegant.  Ideal solution would have hooked into routing and somehow listened to the routing events, updating the DummyComponent template information dynamically.  After over 10 hours sunk into trying to figure that out, I went with the dynamic creation of DummyComponent and their templates instead, which renders a simple component template with the route path text.