Using mlConfigPaths and only one custom command for database management
This allows for mlDeployDatabases to work. It also removes some custom commands that aren't needed if we rely on ml-app-deployer's support for ordering databases based on their dependencies on triggers/schema databases. 

This depends on ml-app-deployer 3.10.1, which adds a couple enhancements to allow for DHF to remove most of the custom commands. 

This also removes some unused methods.