Better fix for passing Gradle properties into HubConfigImpl
Added a comment for the part I'm not sure about - given that we don't want HubConfigImpl to load from gradle.properties when using Gradle (since Gradle has already loaded all the properties from that file and applied the properties plugin to them), when do we want to do this? Does the QS app expect to read from a gradle.properties file? Maybe just JUnit test cases?