1546 Using DeployOtherDatabasesCommand to deploy all databases
The set of staging commands no longer has the database commands; since the final commands are executed first, they have the database commands, and DeployOtherDatabasesCommand is modified to use a custom impl of DeployDatabaseCommand that accounts for some DHF-specific stuff. 

Will soon have a single list of commands a single AppConfig.