Fixing hubInit when gradle.properties doesn't exist (#1632)
* Fixing hubInit when gradle.properties doesn't exist

This requires the user to run "hubInit" if the project hasn't been initialized yet. 

Also, initHubProject is no longer called if the project has not been initialized. If the project hasn't been initialized and the user isn't running hubInit, then the refreshProject call would likely fail because gradle.properties likely isn't present. 

Also removed the logging of confir dirs, that's done by HubAppDeployer. Only had that in for debugging initially.

* Let's actually inithub before we start this these tests