DHFPROD-2194: fix new step dialog validation
 - required asterisks for UI added
 - stepType reference
 - Plus fixes to validation & enum use

——————————————————————————————

required attribute for query vs collection fields removed

Adding an asterisk by putting the "required" attribute, even if that property is dynamic, makes both required.  There appears to be something odd about Angular material that once a field is set to required, it can't be changed back.

So, for now, we're using "*" in placeholder text.  Good enough for beta...