DHFPROD-2497: Attempting to use bulk transform in DHF
The document-update-post-transform.xqy and document-model-update-transform.xqy modules are copied as-is from ML 9.0-9.1, with one exception - the imports are modified to point to DHF libraries. internal-version.xqy was copied as well to satisfy a new endpoint in the rewriter. 

This seems to work, but in testing the ingest-only flow in dh-5-example, it was extremely slow - like tens of seconds to ingest hundreds of documents. When I switched back to the develop branch, ingest was back down to a few seconds. So this doesn't seem worth integrating until we can bump up the minimum version of DHF to 9.0-9 and test out bulk transforms then.