DHFPROD-2811: Mapping steps can now perform entity validation
validateEntity:false is now included in scaffolded mapping steps.

Most of the changes are in test modules; just a handful are in main modules.

Commented out a couple lines in tests that were logging data.

Modified a couple tests to ensure they delete the documents that they create. 

Adding data-hub-test-helper.xqy to make it easier to create entities/flows/etc in unit tests. Rewrote a few tests to make use of this.