DHFPROD-3193: Ensuring language is not submitted in a database payload
This fixes the error in UpdateIndexesTaskTest, but the test still fails for me locally for what seems like the following reason - the number of path range indexes isn't increased because hubSaveIndexes is called before the count of current indexes is retrieved. 

I don't think this test is really testing much that isn't covered elsewhere, and for that reason, I think it would suffice for the test to just verify that no errors occur.