Demonstration of using a data service to get a source data document
Note the TODO in the controller code - the GUI should really pass the number shown to the user as opposed to a count. And the GUI would need to be changed to expect a single document to be returned as well. 

This should consistently perform very quickly, as we're only asking ML to find a single URI.