DHFPROD-3159: No longer overwriting ml-config directory during upgrade
Instead, upgradeHub only backs up the hub-internal-config directory. It then checks to see if range-element-index has an empty array in final-database.json, and if so, it's removed without overwriting the file. 

Also removed the support for replacing the DHF version in build.gradle via regex, as our instructions for an upgrade tell the user to do this themselves first. 

Also did a little renaming while I was in DataHubImpl so that "install into DHS" is now "deploy to DHS".