DHFPROD-3407: Empty/missing expressions no longer throw errors
I am not sure if this is the expected behavior, but it seems better to log and ignore these conditions as opposed to throwing errors, which happens via the Test button in QS - though the errors aren't reported, they're just logged. 

Also ensuring that the "map" namespace prefix is visible in the compiled XSLT. I have not been able to reproduce this issue via a unit test yet, but it happens via QS when using a nested entity property.