DHFPROD-3380: Map prefix is guaranteed to exist in mapping stylesheet
This is needed for the stylesheet generated for core.sjs which results in several map:* references that don't always resolve, depending on the context in which es.mapToCanonical is called. This ensures - via some admittedly hacky code in mlGenerateFunctionMetadata - that the map namespace prefix is declared. 

Also removed some dead code from lib.sjs that was causing confusion about mapping validation.