DHFPROD-3176: Preventing error when fullOutput = true
The main cause of the error was that "documents" in the JSON returned by mlRunFlow was an array of objects, not a map. So ResponseHolder.documents is now a list. 
QueryStepRunner then had to be modified to convert this list into a map structure, where the key is the URI. 

This feature is undocumented and does not have any tests, so I put the change in QueryStepRunner into a try/catch to ensure it can never break anything.