DHFPROD-3755: Can now run a flow via an executable jar
Part of this involves taking the logic in RunFlowTask for processing user inputs and tossing that into the new CommandLineFlowInputs class, where it can be easily reused by the client Main program (and can also be tested easily too). 

Otherwise, this is mostly just some new code - the com.marklogic.hub.cli.client package, and then tests for the new code.