DHFPROD-3974: HubConfigImpl no longer has to refresh Spring-managed beans
This allows for the new UI project to not deal with HubConfigImpl, which won't be a singleton, having dependencies on singletons managed by Spring. 

In addition, the classes that were extending ResourceManager and being managed by Spring no longer extend ResourceManager. Instead, they create a subclass of ResourceManager when needed, based on the current settings of HubConfig.