DHFPROD-4071: Failure to generate search options no longer fails loading modules
The ticket for this didn't specify how search options failed to be generated, only that "{}" was written to an XML file, which of course caused a problem when the options were deployed.

Now, any error occurring while generating search options is logged with a stacktrace. And if an error occurs, the search options files won't be updated.