DHFPROD-4079: dhsDeploy now deploys additional permitted resources
Moved the guts of this from DataHubImpl to DhsDeployer. The "dhsDeploy" method in the DataHub interface was only used by the DhsDeployTask and thus is safe to remove (it is really an internal API that never should have been exposed in the DataHub interface). 

Removed LoadHubSchemasCommand, which has been OBE for multiple releases. 

Added some constructors to DHF objects so that they can more easily be reused in a non-Spring context, which is the case for DhsDeployer.