DHFPROD-4142: Ingestion step can now use relative path without HubProject
Removed FileCollector's dependency on HubConfig - it didn't need an entire HubConfig, it just needed an already-resolved inputFilePath. WriteStepRunner now has the logic for resolving the inputFilePath. 

And to make it easier to unit test WriteStepRunner, moved the instantiation of a DatabaseClient to right where it's needed, if it's not null already.