DHFPROD-4158: Can now deploy to DHS via multiple roles
Most of the files in here are for a new example project - dhs-example. 

Note that "dhsDeploy" is being retained, but as an alias to "hubDeploy". The prior functionality of dhsDeploy - only loading modules and updating indexes - was intentionally replaced. But to prevent errors from the task no longer existing, it's aliased to hubDeploy.