DHFPROD-3865: Removing dead code from mlRunFlow
FlowResource in QueryStepRunner is the only thing that calls this endpoint, and it never passes a "uri" param, nor does it ever send a request body. Also, our docs for using the REST API don't document either of these options.

I'm removing this code in preparation for sharing code between this endpoint and the upcoming DS endpoint for running steps. Wanted to only reuse what's actually needed.