DHFPROD-3854: sourceQuery can now be a script
Step option "sourceQueryIsScript:true" can be added so that the collector knows to evaluate the sourceQuery as a script instead of tossing it into a cts.uris call. 

Did some renaming of "uris" as a variable to "items". Unfortunately can't change e.g. "uris" in a Batch document, as that'd be a breaking change.