DHFPROD-4162: Initial design for managing entity types
This is intended for a design review for 4162 before we start on stories. 

This includes a new tool - the Gradle task generateFilesForDataServices - that will automatically create DS files for each endpoint under ./specs/reference/data-services. 

In addition, Gradle tasks for generating Java interfaces are now dynamically created so that we don't need to update build.gradle every time we add a new service.json file.