DHFPROD-3841: Can now reference $URI in mappings
This opens the door to all sorts of variables in a mapping artifact. However, to achieve this, I had to make a copy of the ES functions map-to-canonical and node-map-to-canonical so that I can pass in a map of params to the underlying xdmp:xslt calls. I opened up bug 54632 so that the ES functions will allow this in a future version.