DHFPROD-4752: Generating a swagger-ui view for the models OpenAPI file
Using JsonRefs to resolve references to the models directory, but it unfortunately does not yet resolve references between model files.

This adds a new route at "/", which serves up index.html, which provides links to each swagger-ui view. So the existing /api/swagger-ui/ view is not modified in any way.