DHFPROD-4765: Fixing nightly test failures
Deleting EnvironmentControllerInstallTest, as we know that install/uninstall have to be removed.

Removed two failing tests from DataHubInstallTest - we have these covered already in the hub.core.DataHubInstallTest, they seem to be complete duplicates. The two in this test were failing because they're creating a new version of the DH test app and not running the code in the Installer program that adds the "status" privilege to the flow-developer user.

Removed the mlDeploy/mlUndeploy from CreateHarmonizeFlowTaskTest and moved it out of fullcycle. The test ran fine for me locally.