DHFPROD-4440: Improving how resetProject works in HubTestBase
Main change is that clearDatabases in HubTestBase no longer deletes the hub-core-artifact collection. That keeps all DH core artifacts present. Had to improve some legacy tests that were assuming that the database was empty after calling clear (making assertions on the number of docs in a database leads to brittle tests).

Also made a tweak to LoadUserModulesCommand that both speeds up tests and fixes a long-standing issue when running "mlWatch", which is - when you run mlWatch, DHF generates and deploys entity-based search options over and over and over. That also adds some time to every test, and none of our tests depend on LoadUserModulesCommand deploying entity-based search options.