DHFPROD-4751: Reworked HC to default to connecting to DHS
The new HubCentral class is now responsible for collecting all configuration and helping configure a new HubConfigSession when a user logs in. This makes it easy to determine whether to use the new defaults for DHS or not. 

In the process, I removed a lot of the stuff pertaining to features that we know are either being removed (installing DHF) or being redesigned (upload/download/reset). 

Also bumped up the bugfix version numbers of our build.gradle file and simplified the set of dependencies. 

The README has been updated to describe how to run HC locally now. 

Finally, did a little more cleanup from the PR for renaming one-ui to HubCentral - I renamed AbstractOneUiTest, and I moved all the curation.controllers under central.controllers.