DHFPROD-4440: Steps now support custom processors
Included a new "step-processors" example project to document the feature. 

There are a lot of changes in flow.sjs, but much of this was due to 1) renaming variables for consistency, and 2) moving existing code into two new functions - writeProvenanceData and updateBatchDocument - so that "runFlow" isn't quite so big. 

The actual new stuff in flow.sjs is fairly minimal - it's the new applyProcessorsBeforeContentPersisted function, and I had to change the if/else that invokes a step module based on acceptsBatch=true. That block of code no longer adds content objects to the writeQueue, because it first needs to apply processors before adding content to the writeQueue.