DHFPROD-4720: Can now download user configuration files
Added an "all-artifacts" project in core to support this. 

Added our first use of "Secured" in HC, along with a pattern for how to verify that a user can or cannot access a Secured controller method. 

Also removed a number of unused classes that aren't needed now that we're not checking if the user can install DH - ForbiddenException(LoginFailureHandler can assume that any failure is due to the user being unauthorized); ProjectDirectoryException (no more project); UIDeployListener (no more install); DataHubProjectUtils (no more upload). Also removed BadRequestException - CustomExceptionHandler wasn't doing anything with it, and it was being improperly thrown by MapSearchManager instead of the generic DataHubException (it shouldn't have been thrown there, since it was extending AuthenticationException).