DHFPROD-4757: Removing HC dependency on the filesystem
Added a BaseController that is the one and only one bean that accesses HubConfigSession. Every controller extends that and is now abstracted from how a HubConfig is obtained.

Removed every controller method that both did not have a test and that I could not find any way to access it via the UI code. Doing this to bump up code coverage.

Renamed methods in subclasses of AbstractArtifactController and changed the tests to call those methods; again, to improve code coverage. 

Changes to core are minimal - removed file-writing from ArtifactManagerImpl; added a null-check to HubProjectImpl so it doesn't assume a token exists (this doesn't impact HC app code, but HC test code - which does need to initialize project - will otherwise break without this check); and made a couple methods in EntityManagerImpl and FlowManagerImpl visible to HC (the methods have useful logic that don't depend on a filesystem). 

Removed two deleted endpoints from the mocks files.