DHFPROD-4775: New endpoint for clearing user data
The interesting new stuff is all in DataHubImpl.

Tossed a bunch of the download-testing stuff into AllArtifactsProject so that it can be reused easily by ClearUserDataTest.

Modified the "runAs*" methods so that the ManageClient runs as the given user too to avoid confusion. This caused a number of existing tests to break, as they were assuming that the ManageClient was still associated with an admin or flow-developer. So I fixed up those tests to ensure they're running as the proper user based on the operation being performed.