DHFPROD-4887: Generating Swagger docs dynamically via springfox
Added jsonSchema2Pojo Gradle task to generate Java classes from ./specs/models. Where possible, these are being used for documentation purposes in controllers. I renamed some of these to not have ".v1" in their name, so e.g. "ModelDescriptorV1" seemed like a funky class name. I imagine we'll get rid of ".v1" in all the filenames soon, and replace it either with ".schema" or with nothing.

The README is updated to point to the new endpoint for Swagger docs. It also points to the Application class for more information on how to configure annotations on endpoints. Existing endpoints are a good source of info as well.

I tried to document every endpoint possible; some are using anyObject in mocks.json today, so I didn't make Java classes for some of those. We can gradually fill in those gaps.

If all looks good, I'll submit a separate PR to remove the existing mocks support in the ./ui folder.