DHFPROD-4865: primaryEntityTypes endpoint now includes entityTypeId
Removed Address/Zip entity models from the reference project, as we don't want to support that approach anymore.

With this change, mappings no longer appear in the UI because the UI is expecting targetEntityType to be "Customer". But the backend does not support that; entityTypeId must be used instead. So the UI now needs to:

1. Create mappings with targetEntityType = entityTypeId
2. Find mappings for an entity type based on entityTypeId, not entityName

Updated getArtifactsGroupByEntity in core.sjs to support querying on either entityNames or entityTypeIds.