DHFPROD-4963: ML unit tests can now reset databases before each suite
This means a test doesn't have to clean up after itself, which makes debugging easier - if a test fails, you can fiddle with the results left in the database. This mirrors how many of our JUnit tests now work. 

In addition, load-artifacts/entities now writes to both staging and final. 

Ideally in the near future, these functions will use the artifacts library for writing artifacts so we get the proper permissions too. 

Modified RunMarkLogicUnitTestsTest to reset the databases before and after all the tests run too.