DHFPROD-4914: Adding support for hubDhs and hubSsl properties
Reworked/simplified HubConfigImpl so that refreshProject / loadConfigurationFromProperties go through applyProperties now, ensuring that the propertyConsumerMap is the only thing that applies properties, and does so in a consistent fashion. 

One tricky thing was ensuring that DataHubPlugin uses the *Config objects constructed by ml-gradle. I think how that works is now more clear in the code. 

Most of the changes in here are to HubConfigImpl, and most of the changes involved deleting code. 

A number of test classes were updated as well based on some updated method signatures in HubConfigImpl. The HubConfig interface itself has not changed.