DHFPROD-4671: Fixing LoadHubModulesCommandTest
I removed the call to load hub modules, which fails intermittently in the Jenkins environment. And it was unnecessary - we know the hub modules have been loaded already, the point of the test was to verify that the versions match. 

This likely isn't related to 4671, just borrowing that JIRA ID because it's the latest commit on develop right now.