DHFPROD-4857: Can now apply a project zip to a project
Refactored some existing HC-specific stuff in DH core into a HubCentralManager class. 

No longer including step definitions in the downloaded zip. 

Added an initial "deleteModel" endpoint for the sake of testing this feature, but it's not yet ready to be used by the app (future story exists for that). 

So the guts of this are in HubCentralManager and ApplyDownloadedZipToProjectTest.