DHFPROD-5178: Upgrading to Java Client 5.2.0
Upgraded Gradle wrappers to 6.2.2, which also depends on kotlin-stdlib 1.3.61. 

Upgrading marklogic-data-hub and web to use the com.github.node-gradle.node plugin instead of the old com.moowork.node plugin. This is because if you clear your Gradle cache and try to build either project, it won't work - you'll get an error of "Could not find org.nodejs:node:8.10.0". That's presumably due to moowork being out-of-date. The error does not happen with the more recent plugin. 

Made fixes to ml-data-hub Gradle tasks to avoid validation warnings that are now causing failures since Gradle 6 is being used to build the plugin. These were all simple fixes - either adding getters for @Input properties (which Gradle 6 requires), or extending the new AbstractConfirmableHubTask to avoid the same warnings from ml-gradle's MarkLogicTask class. 

Also had to add "@Optional" to every "@Input" property or otherwise Gradle complains. 

Made some fixes to the QuickStart classpath as well so that the proper versions of kotlin and okhttp3 are used. 

Deleted unused, unreferenced Nifi task.