DHFPROD-5286: Processing URIs for a step now uses DS endpoint
All inputs are passed in a JSON object to the DS endpoint. This fixes an issue where the mlRunFlow endpoint fails when the string of request parameters is too long (mostly due to the length of each URI and the batch size). 

Cleaned up a few other things while I was at it:

- The sourceDatabase field wasn't used in QueryStepRunner, so removed it from QSR and from StepRunnerFactory (as it doesn't apply to WriteStepRunner either)
- The destinationDatabase field was actually never used in QSR either; the "target-database" param passed by FlowResource to mlRunFlow was never used. 
- Modified StepRunnerFactory to only set destinationDatabase on WriteStepRunner and removed it from the StepRunner interface