DHFPROD-5393: Can now run hubUpdate on DHF 4.3.x
Added some logging and a try/catch on updating mapping steps, just in case the call to get the ML version bombs. 

The actual fix is in Versions - we now fallback to ml:hubversion if mlHubversion can't be hit. See the comments about the complexity/value of an automated test for this.

Finally, I bumped up the Gradle version in the scaffolded gradle-wrapper.properties to be 6.2.2. The user will really need to do this first (since it can't be done as part of "./gradlew hubUpdate), but I wanted to correct this here as I think not doing it was an oversight in the ticket for upgrading to Java Client 5.2.0.