DHFPROD-5259: Securing DS endpoints with privileges
Most changes in here are just one-line additions to DS endpoints. 

Did some cleanup - the jobInfo/getLatestJobData endpoint is now part of ModelService, as it's a model-specific function. And the name "JobInfo" was inconsistent with all other service names. 

Added the data-hub-write-custom privilege so that endpoints for modifying custom steps could be secured. Only a data-hub-developer has the privilege, as no HC user can do this yet.