DHFPROD-5635: Adjusting fix for what steps to download
This is intended to avoid downloading future step types that we support - i.e. matching, merging, mastering - without having to adjust the code. Once HC is able to modify another step type in the future, we'll modify this to account for that. 

No test changes, just made sure the existing tests for downloading steps still pass.