DHFPROD-5626: Fix for facet-values APIs
The es namespace and entity def namespace are now included when calling cts.pathReference.

I am not convinced these are fully working yet - the cts.pathReference call may need the second argument populated to specify the correct datatype and collation. Going to do some more testing, but wanted to get this merged ASAP as it's breaking HC. 

Also fixing GenerateProtectedPathsTest on 9.0-11 - need to use DeployHubQueryRolesetsCommand, which ignores the bug in the Manage API that throws an error when updating a query roleset.