DHFPROD-5638: Download and clear-user-data roles are now limited
The download role didn't need data-hub-operator, and the tests for it still ran fine.

The clear-user-data role no longer needs data-hub-developer, but I was not able to get it work - it kept failing on writing documents back - until I added data-hub-operator, which presumably adds a privilege that is needed. This seems like a reasonable role for this role to inherit, as clearing user data seems more powerful than data-hub-operator.