DHFPROD-5712: Removing hubGenerateExplorerOptions task
I can't figure out if there was ever a reason for this, as saveQueryOptions was always generating explorer options and deployQueryOptions was always deploying them.

Also simplified EntityManagerImpl by getting rid of the timestamp check. In every scenario where deployQueryOptions is called, we want to regenerate the options based on entity models - it's very fast, and it eliminates any chance of options being out of sync with the models. This allowed for HubModuleManager to be deleted as well, which was only used by EntityManagerImpl.