DHFPROD-5768: Gradle task groups now align with what docs will be
I moved a lot of task definitions around so that the groups are all together. To verify no tasks were lost, I temporarily added this line:

println "TASK COUNT: " + project.tasks.size()

before and after the tasks are added. Before and after this ticket, the count of tasks before the DHF plugin was applied was 170, and was 213 afterwards. So 43 tasks are still being added (just wanted to verify I didn't drop any tasks). 

Also deleted ExportJobsTask, as it wasn't used anywhere.