DHFPROD-5840: Avoiding null pointer while upgrading project
Without changing the code quite a bit, it's not possible to write a test for this condition because the null pointer exception (which occurs when stepDefinitionType is not set, which is true for converted steps) is caught and logged. And it's the logged error message that's confusing for users. So this just modifies the "if" condition so that the values being tested are on the right hand side of "equals" instead of the left hand side, thus avoiding the chance of a confusing NPE.