DHFPROD-5811: Fixing failing tests
The use of newAppServicesDatabaseClient is failing intermittently. Since we don't need to retain any documents, we can use the Manage API to clear the database. 

Also added some primitive retry support for resetDatabases. 

Modified DeployUserAmpsTest to only run if nightly build has major version of 10. Also added a little unit test to verify that "major" is set if it's a nightly build.