DHFPROD-5883: Adding tests for connector
Main change here is to assume the use of DHF-supported properties that are passed in the params map to the DataWriterFactory. Those properties can then be used to construct a HubClient for communicating with DHF. 

Bringing in the BulkIngest work too so we can exercise that as part of the test.