DHFPROD-4854: Removed spring-boot dependency from core project
For the core project, we only need spring-context, and we now obtain a ConfigurableApplicationContext where needed by just instantiating an AnnotationConfigurableApplicationContext with the appropriate Configuration class. 

Replaced the spring-boot bootJar task with a custom jar task that is almost identical to the existing one for clientJar - just has a different main class. 

Renamed logback-spring.xml to logback.xml, and this no longer is included in the DHF library jar, since processResources no longer copies it in. It's only included in the executable jars that need it.

Removed EnableAutoConfiguration from DataHubPlugin, which wasn't even doing anything. The Spring container is created manually in that class. 

Modified EntityModelConverterTest to no longer test an empty JSON file. The most recent version of Jackson treats this as a MissingNode instead of an ObjectNode. But technically, an empty JSON file is invalid and should throw an error if it exists in a user's project.