DHFPROD-5945: Schema for a DH envelope
This is being added to the develop branch so that it can be used by the Spark connector team without having to switch to the 5.4-develop branch. It has no impact on anything in 5.3, other than if the Gradle task generateSchemaPojos is run. And I verified that it runs successfully.