DHFPROD-6116: Fixing QS on 5.4-develop
WebApplication was ending up with 2 HubConfigImpl instances due to ComponentScan'ing all of com.marklogic.hub. I was going to just mark the HubConfigImpl in WebApplication as @Primary, but figured it's safer / easier-to-understand if the packages are listed out. 

Also fixed an issue in CurrentProjectController, which was autowiring in HubProject. Added a simple test to ensure that this class can be autowired successfully.