DHFPROD-6113: Removing need to include skipUi or skipWeb
The UI files for core will only be built when running "publish" and "build" now. 

Same goes for HC, except only when "bootWar" is run. Removed the maven-publish plugin from this project too, as it is not needed. 

skipWeb is no longer needed as QS will only publish something when "publish" is run. Removed publishing from the web project too; confirmed with team it's no longer needed. 

Removed the Spark README, as the only thing it was doing was mentioning the need to include skipui. We'll add it back once we have stuff to say in it. Also removed publishing from this, as we have no use case for that yet.