DHFPROD-6096: Spark connector now handles write failures and aborts
Also removed the concept of batchSize - the BulkInputCaller already supports this itself, we don't need to maintain a list of records in HubDataWriter. 

As a result of this, rewrote a lot of the tests so that HubDataWriter.commit is called after all the fruit rows are written. This ensures that any pending writes are flushed.