DHFPROD-6160: Repackaged writer tests
Also extracted HubDataSourceWriter so that it's in the "writer" package. This sets us up nicely for the upcoming Reader stuff.

Fixed a couple tests too so that they run on DHF 5.2.x - changes were just to the test modules, not the real stuff. 

Updated the project README as well, and renamed the two test programs for clarity.