DHFPROD-5997: Fixing logic for determining step success
The problem was that CollectorImpl was not throwing an exception when the collector failed. It instead returned a single item - the error message (the serialized JSON object from the REST API). That was then processed by the step, which resulted in zero successful items - but also zero failed items. 

Now that the collector properly throws an error, QueryStepRunner can say that if there are no failed items, then the step completed successfully.