DHFPROD-5997: Fix for determining failed batches
Made what I think is a fix to QueryStepRunner so that if the count of failed items and total items in a batch are both zero, then the batch is considered successful. A batch is now only considered to have failed if there's at least one item that threw an error, which seems to make good sense.