DHFPROD-6412: Fixing syntax error when writing prov data
Also improved ML unit test plumbing by making "clear-jobs-database" delete provenance data as well, which depends on an amp. 

One other change to flow.sjs is that previously, a bunch of prov data could be generated, but if the granularity level is off, then nothing was done with it. It seems to make more sense that if the level is off, we can bail right away and not generate all that data.