DHFPROD-6476: Step processor failure now reports correct job counts
Had to remove the throwing of an exception from stopJob - I couldn't find any reason why that would be good. With it there, the test was failing because 4 failed items were being reported instead of 2. Specifically, a stepItemFailureListener was throwing this exception, after the 2 failed items had already been captured. So it seems unnecessary to throw an exception when stopJob is called but the jobId cannot be found. Also ran all of the flow tests and none of them failed with this change.