DHFPROD-6527: No longer generating empty range-element-index array
I looked into the history of this change, and I think when it was made for DHFPROD-6215, the intent was good - try to remove indexes that previously existed - but it wasn't being done consistently for path range indexes as well. Nor is removal being done anywhere else - e.g. for protected paths based on PII properties. But we know that an empty array can be a dangerous thing, as that can remove all indexes from a database. And in the context of the entity-config database files, there's no way to know if all indexes should be removed or not. So an empty array is no longer being generated.