DHFPROD-6560: mlWatch now generates function metadata
I tried to get a test in place for this, but I don't know how to test mlWatch in a test (which enters an infinite loop). Instead, I added some logging to at least indicate when there's a misconfiguration -  i.e. the command or context are null. 

The actual bug fix is just to populate generateFunctionMetadataCommand when the other beans are populated in DataHubPlugin.