DHFPROD-6629: Fixing intermittent test failures
Converted jsonToXmlIngestion into a JUnit test since we don't know what's going on 
with that. 

Made a fix to CustomStepE2E to ensure it doesn't impact ML unit tests.

Lowered expectation in MasterTest.

Fixed asymmetric test by making it not return the URI of the entity model that is saved.