DHFPROD-6674: Removing DS prototype for running a step
This was created in Feb 2020, but there are no plans for it to be used anytime soon. Removing to avoid any confusion over whether it should be used. It introduced two new params - contentDescriptorLimit and a filterQuery for findMatchingContent - that are not needed anywhere else, so those are removed too.