DHFPROD-7037: New junit5 subproject
Most changes are in the new junit5 subproject.

As part of the new project depending on the latest JUnit release (5.7.1), I did a lot of cleanup in all of our Gradle files so that DHF uses 5.7.1 across the board. Also deleted Gradle config - like source and target compatibility - that is already being set at the root project level.