DHFPROD-6526: Fixing job times bug via new DS endpoints
The mlJobs endpoint was trying to handle every job operation, and I was having a hard time sorting out how to fix it so that the timestamps were always correct and valid on the Job document. So I replaced it with 4 new DS endpoints and a JobService interface, which I think is a lot easier to understand. And the "RunFlowWithSameJobIdTest" is now passing, which I updated to include multiple steps, as that produced some new problems that  the original fix wasn't handling. 

I removed all the unused stuff as a result of this change - that includes JobDocManager, JobMonitor, and a number of functions in jobs.sjs. 

Also, jobs.sjs is now a library module, which makes it a bit easier to use.