DHFPROD-6718: Connected steps now generate job data
Most of the changes here involve moving logic from larger library modules (jobs.sjs and flowRunner.sjs) into smaller class modules, which should make the logic easier to understand and result in a better separation of concerns. 

The main new chunk of behavior is the addition of the "stepResults" array to the Batch document. 

Added Job.sjs and Batch.sjs classes to consolidate logic pertaining to creating/updating these documents.

Added FlowExecutionContext.sjs to capture state pertaining to a flow that is running connected steps.