DHFPROD-7114: Cleaning up custom step template and starting public API
Trimmed down main-custom.sjs to focus on what the inputs and outputs are, along with a few simple things that can be done in conjunction with makeEnvelope. 

Taking a stab at a true public API in DHF/ML code as well via flowApi.sjs. 

Removing scaffold templates for mapping/mastering - those are OBE since QS is gone.

Not touching custom-ingestion.sjs yet - I think the value of that template is extremely low since ingestion steps don't have much value in the first place.